Showing only actionable tags when selecting item tags
